refactor(channels): adopt future-rpc for proto, decode and event parsing
Batch 8 of the typed-RPC milestone: the channel bridge becomes the first
client on the shared wire contract.

- grpc_client.rs consumes future_rpc::proto (re-exported under the
  historical module name); the checked-in channels/src/generated/proto.rs
  copy is deleted and build.rs no longer codegens future.proto (feishu_ws
  stays, so tonic-build drops out of the build-deps).
- call() decodes through future_rpc::decode::response_data — typed payload
  first, JSON `data` fallback — so every command reads identically against
  old and new agents.
- AgentClient::parse_event now delegates to future_rpc::events::
  parse_agent_event; the AgentEvent enum and its parser moved into the
  crate so every future client decodes events the same way. The parser is
  typed-first with the JSON fallback; tool_args keeps its wire semantics
  (string-form only) on both paths.
- expand_projection_snapshot carries the typed payload through, so
  reconnected/replayed events decode through the same path as live ones.
- Parse tests moved to the crate, where each case now runs as twins: the
  JSON-data event (old agent) and the typed-payload event (new agent).
  Channels keeps projection-expansion, SessionState and ImageInput tests,
  plus a delegation smoke test on both wire states.
- The (ignored, live-agent) integration test switches to future_rpc proto
  + decode instead of its private generated copy.
